프로젝트 api 수정 - 변경된 디비, 특정 유저, 그외 에러
