--- applications/projects/controllers/projectListController.js
@@ -1,15 +1,0 @@
-const knex = require("../../../db/knex");
-const { select } = require("../../../db/knex");
-
-let projectListController = {};
-
-projectListController.read = function (req, res) {
-    knex("project")
-        .select("projectName", "projectId")
-        .where("userId", res.locals.userId)
-        .then((projectList) => {
-        res.json(projectList);
-    });
-};
-
-module.exports = projectListController;
